+++ src/runtime/ext/ext_network.h
@@ -87,11 +87,11 @@ bool f_headers_sent(Variant file = null, Variant line = null);
 
 void f_header_remove(CStrRef name = null_string);
 
-bool f_setcookie(CStrRef name, CStrRef value = null_string, int expire = 0,
+bool f_setcookie(CStrRef name, CStrRef value = null_string, int64 expire = 0,
                  CStrRef path = null_string, CStrRef domain = null_string,
                  bool secure = false, bool httponly = false);
 
-bool f_setrawcookie(CStrRef name, CStrRef value = null_string, int expire = 0,
+bool f_setrawcookie(CStrRef name, CStrRef value = null_string, int64 expire = 0,
                     CStrRef path = null_string, CStrRef domain = null_string,
                     bool secure = false, bool httponly = false);
